[APP-865] replace mock data with real call to our BE.

+++ feature_apps/src/main/java/cm/aptoide/pt/feature_apps/domain/Bundle.kt
@@ -11,7 +11,7 @@ data class EditorialBundle(
   val image: String,
   val subtype: ArticleType,
   val date: String,
-  val views: Int,
+  val views: Long,
 ) :
   Bundle(editorialTitle, emptyList(), Type.EDITORIAL)
