+++ feature_apps/src/main/java/cm/aptoide/pt/feature_editorial/presentation/EditorialView.kt
@@ -21,7 +21,7 @@ fun EditorialView(
   subtype: ArticleType,
   summary: String,
   date: String,
-  views: Int,
+  views: Long,
 ) {
   Column(
     modifier = Modifier
